split event manager into own package
Move the event manager into a package of its own. This is easily done
because the event manager has no dependencies on any other uzbl
components, since all it needs is a stream of text and a list of
handlers to call.

This change does, however, necessitate some changes to the
initialization process and "extension" handling. Because the EM won't be
available until an actual uzbl instance has been created, extensions
cannot register their handlers before this point. That's why a new
interface, Registerable, has been added, as well as a Register()
function to register extensions. Said interface consists of a single
Init() method, that will be called for all extensions once the EM (and,
by extension, the input system) are available.